refactor(core): move clip option helpers out of types
