+++ packages/core/src/types/rendering.d.ts
@@ -1,5 +1,4 @@
 import Mark from "../marks/mark.js";
-import { MarkProps } from "../spec/mark.js";
 import { LocSize } from "../view/layout/flexLayout.js";
 import Rectangle from "../view/layout/rectangle.js";
 import ViewRenderingContext from "../view/renderingContext/viewRenderingContext.js";
@@ -74,37 +73,6 @@ export interface BufferedRenderingRequest {
     coords: Rectangle;
     clip?: ClipOptions;
 }
-
-export function normalizeClipOptions(
-    options: RenderingOptions
-): ClipOptions | undefined;
-
-export function createClipOptions(
-    rect: Rectangle,
-    clipX: boolean,
-    clipY: boolean
-): ClipOptions | undefined;
-
-export function clipCoords(
-    coords: Rectangle,
-    clip: ClipOptions | undefined
-): Rectangle;
-
-export function combineClipOptions(
-    current: ClipOptions | undefined,
-    next: ClipOptions | undefined
-): ClipOptions | undefined;
-
-export function createSelfClipOptions(
-    clip: MarkProps["clip"],
-    coords: Rectangle
-): ClipOptions | undefined;
-
-export function prepareMarkClipOptions(
-    options: RenderingOptions,
-    markClip: MarkProps["clip"],
-    coords: Rectangle
-): ClipOptions | undefined;
 
 /**
  * Method signature for View.render(). Reduces the amount of JSDoc needed.
